move 全ての SKNode 向けのモディファイアを Commons に移動しました.
