+++ Sources/GameWidget/Commons/Modifiers/ForAllNode.swift
@@ -1,8 +1,8 @@
 //
-//  NodeModifiers.swift
+//  ForAllNode.swift
 //  
 //
-//  Created by rrbox on 2022/10/30.
+//  Created by rrbox on 2022/11/28.
 //
 
 import CoreGraphics
@@ -124,22 +124,5 @@ public struct ZPosition<Context: ZPositionContextProtocol>: Modifier {
 public extension ModifiableWidget where Context: ZPositionContextProtocol {
     func zPosition(_ value: CGFloat) -> Next<ZPosition<Context>> {
         self.modifier(mod: ZPosition(value: value))
-    }
-}
-
-
-public struct NodeContext: PositionContextProtocol,
-                           RotatableContextProtocol,
-                           ScalableContextProtocol,
-                           ZPositionContextProtocol {
-    
-    public var position: CGPoint = .zero
-    public var zRotation: CGFloat = .zero
-    public var xScale: CGFloat = 1
-    public var yScale: CGFloat = 1
-    public var zPosition: CGFloat = .zero
-    
-    public init() {
-        
     }
 }
